+++ web-components/src/components/cards/TebuCard/TebuCard.Duration.tsx
@@ -3,6 +3,7 @@ import cn from 'classnames';
 
 import InfoIconOutlined from '../../icons/InfoIconOutlined';
 import InfoTooltip from '../../tooltip/InfoTooltip';
+import DashedLine from './TebuCard.Duration.DashedLine';
 
 interface TebuCardDurationProps {
   duration: number;
@@ -15,31 +16,7 @@ interface TebuCardDurationProps {
   maximumPrefix?: string;
 }
 
-interface DashProps {
-  style?: React.CSSProperties;
-  className?: string;
-  height: number;
-}
 
-const DashedLine: React.FC<DashProps> = ({ style, className, height }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="2"
-    height={height}
-    viewBox={`0 0 2 ${height}`}
-    fill="none"
-    className={className}
-    style={style}
-  >
-    <path
-      d="M1 0L0.999999 26"
-      stroke-dasharray="4 4"
-      style={{
-        stroke: 'rgba(var(--ac-neutral-500) / 1)',
-      }}
-    />
-  </svg>
-);
 
 const TebuCardDuration: React.FC<TebuCardDurationProps> = ({
   title,
