+++ web-components/src/components/cards/TebuCard/TebuCard.tsx
@@ -4,6 +4,7 @@ import Card from '../../cards/Card';
 import InfoIconOutlined from '../../icons/InfoIconOutlined';
 import InfoTooltip from '../../tooltip/InfoTooltip';
 import { Body, Label } from '../../typography';
+import cn from 'classnames'
 
 export interface TebuCardProps {
   className?: string;
@@ -24,8 +25,8 @@ export default function TebuCard({
   headerTooltip,
 }: TebuCardProps): JSX.Element {
   return (
-    <Card className={`text-sc-text-header w-[310px] ${className}`}>
-      <Box className="flex items-center gap-[10px] pl-[20px] pr-[1px] py-[15px] text-[12px] bg-sc-card-standard-header-background">
+    <Card className={cn('text-sc-text-header w-[310px]', className)}>
+      <Box className="flex items-center gap-[10px] pl-[20px] pr-[5px] py-[15px] text-[12px] bg-sc-card-standard-header-background">
         <Body className="font-montserrat tracking-[1px] text-[12px] font-montserrat uppercase font-extrabold text-sc-text-sub-header">
           {header}
         </Body>
